Dataset Information
DATASET INFORMATION :

The dataset consist of 7350 observations. The features of the dataset are: customer_id: Every customer is given a unique ID gender: Whether the customer is a male or a female location: Location of the customer partner: Whether the customer has a partner or not (1=Yes, 0=No) dependents: Whether the customer has dependents or not (1=Yes, 0=No) senior: Whether the customer is a senior citizen or not (1=Yes, 0=No) Tenure: Number of years the customer has stayed with the company. Lu’s communication offers a loyalty scheme; this scheme gives 2% discount on the monthly cost for each year (up to 25 years) the customer remains with the company monthly_cost: The amount charged to the customer monthly package: Packages Lu’s communication offers. See table below for more information survey: Score given by customers on the customer service (0=Poor”, 10=Excellent”) Class: Whether the customer churned or not